==> Automating Equity Project/2. Cleaning and Preprocessing/1. language_detection_preprocessing.ipynb <==
# ATTN: This script uses Google translate to detect job description language. Google translate will limit requests and take a very long time. Only run this script if redoing language detection.

# Read from scraped data

import os # type:ignore # isort:skip # fmt:skip # noqa # nopep8
import sys # type:ignore # isort:skip # fmt:skip # noqa # nopep8
from pathlib import Path # type:ignore # isort:skip # fmt:skip # noqa # nopep8

mod = sys.modules[__name__]

code_dir = None
code_dir_name = 'Code'
unwanted_subdir_name = 'Analysis'

if code_dir_name not in str(Path.cwd()).split('/')[-1]:
    for _ in range(5):

        parent_path = str(Path.cwd().parents[_]).split('/')[-1]

        if (code_dir_name in parent_path) and (unwanted_subdir_name not in parent_path):

            code_dir = str(Path.cwd().parents[_])

            if code_dir is not None:
                break
else:
    code_dir = str(Path.cwd())
sys.path.append(code_dir)

# %load_ext autoreload
# %autoreload 2


from setup_module.imports import * # type:ignore # isort:skip # fmt:skip # noqa # nopep8


# This is a manually collected dictionary of incorrect/faulty keywords in scraped site data
with open(f'{scraped_data}CBS/Data/keyword_trans_dict.txt') as f:
    keyword_trans_dict = json.load(f)


# 114 words to fix
len(keyword_trans_dict)


def fix_broken_linkedin_files(glob_path):
    fix_list = []
    data_dict = {}
    data_list = []

    if glob_path.endswith('.json'):

        with open(glob_path, encoding = 'utf-8') as csv_file_handler:
            csv_reader = csv.DictReader(csv_file_handler)

            for rows in csv_reader:
                first_key = str(list(rows.keys())[0])
                key = rows[first_key]
                data_dict[key] = rows

        for num in data_dict:
            data_list.append(data_dict[num])

        with open(glob_path, 'w', encoding = 'utf-8') as json_file_handler:
            json_file_handler.write(json.dumps(data_list, indent = 4))

    return data_list


def fix_keywords(df_temp):

    # This is a manually collected dictionary of incorrect/faulty keywords in scraped site data
    with open(f'{scraped_data}CBS/Data/keyword_trans_dict.txt') as f:
        keyword_trans_dict = json.load(f)

    if len(df_temp) > 0 and isinstance(df_temp, pd.DataFrame):
        for key, value in keyword_trans_dict.items():
            df_temp.loc[
                df_temp[df_temp['Search Keyword'].notnull()]['Search Keyword'].astype(str).progress_apply(
                lambda x: x.lower().strip()
                ) == str(key).lower().strip(), 'Search Keyword'
            ] = str(value).lower().strip()

        unfixed = df_temp.loc[
            df_temp[df_temp['Search Keyword'].notnull()]['Search Keyword'].astype(str).progress_apply(lambda x: x.lower().strip()).isin([x.lower().strip() for x in list(keyword_trans_dict.keys())])
        ]

        if len(unfixed) != 0:
            for key, value in keyword_trans_dict.items():
                for idx, row in df_temp.iterrows():
                    if row['Search Keyword'].astype(str).lower().strip() == str(key).lower().strip():
                        df_temp.loc[idx, 'Search Keyword'] = str(value).lower().strip()

        unfixed = df_temp.loc[
                df_temp[df_temp['Search Keyword'].notnull()]['Search Keyword'].astype(str).progress_apply(lambda x: x.lower().strip()).isin([x.lower().strip() for x in list(keyword_trans_dict.keys())])
            ]
        if len(unfixed) != 0:
            print('Some keywords were not fixed. Please check file unfixed_keywords.txt in data directory.')
            with open(f'{data_dir}unfixed_keywords.txt', 'w') as f:
                json.dump(unfixed, f)

    return df_temp


#### Read paths

glob_paths = []

for site in site_list:
    glob_paths.extend(glob.glob(f'{scraped_data}/{site}/Data/*.json')+glob.glob(f'{scraped_data}/{site}/Data/*.csv'))


# 955 json and csv files
len(glob_paths)


#### Use paths to open files, fix keywords, and drop unneeded columns

%%time
# Fix list catches all incorrect/faculty keyword search terms
fix_list = []

# Appended data catches all the fixed and cleaned dfs
appended_data = []

for glob_path in glob_paths:

    if glob_path.endswith('.json'):
        try:
            df_temp = pd.read_json(glob_path)
        except ValueError:
            fix_list.append(glob_path)
            if '1. Scraping/LinkedIn/Data/linkedin_jobs_df_' in glob_path:
                data_json = fix_broken_linkedin_files(glob_path)
                try:
                    df_temp = pd.read_json(glob_path)
                except ValueError:
                    fix_list.append(glob_path)
    elif glob_path.endswith('.csv'):
        df_temp = pd.read_csv(glob_path)
    if len(df_temp) > 0 and isinstance(df_temp, pd.DataFrame):
        df_temp = fix_keywords(df_temp)
        df_temp = df_temp
        # df_temp = df_temp.drop(columns=cols, axis='columns', errors='ignore')
        df_temp = df_temp.drop(
        df_temp.columns[
                df_temp.columns.str.contains(
                    'unnamed|index|level', regex=True, case=False, flags=re.I
                )
            ],
            axis='columns',
            errors='ignore',
        )

        if glob_path.endswith('.json'):
            df_temp.to_json(glob_path, orient='records')
        elif glob_path.endswith('.csv'):
            df_temp.to_csv(glob_path, index=False)

        appended_data.append(df_temp)

# Concatonate list of dfs into one large df_jobs
df_jobs = pd.concat(appended_data, axis='index')

# Save df_jobs to file
if len(df_jobs) > 0 and isinstance(df_jobs, pd.DataFrame):
    df_jobs.to_pickle(f'{df_save_dir}df_raw.pkl')
    df_jobs.to_csv(f'{df_save_dir}df_raw.csv', index=False)
else:
    print(f'ERORR: LENGTH OF DF = {len(df_jobs)}')


# If we couldn't fix some keywords, we add them to list fix_list and write to file
if len(fix_list) != 0:
    print('Some keywords to fix!')
    with open(f'{data_dir}fix_list.txt', 'w') as f:
        json.dump(fix_list, f)


# List of dfs, len = 527
len(appended_data)


# Concatonate list of dfs into one large df_jobs
df_jobs = pd.concat(appended_data)


# len = 204113
len(df_jobs)


df_jobs.columns


# Append keywords to list and save to file
keywords = list(set(df_jobs['Search Keyword'].tolist()))
print(f'Number of unique keywords: {len(keywords)}')
with open(f'{data_dir}unique_search_keywords_len({len(keywords)}).txt', 'w') as f:
    json.dump(keywords, f)


keywords[:10]


# Save df_jobs to file
assert len(df_jobs) > 0 and isinstance(df_jobs, pd.DataFrame), f'ERORR: LENGTH OF DF = {len(df_jobs)}'
df_jobs.to_pickle(f'{df_save_dir}df_raw.pkl')
df_jobs.to_csv(f'{df_save_dir}df_raw.csv', index=False)


# Drop duplicated and missing data

### START HERE IF SOURCING FROM DF_JOBS_RAW
### PLEASE SET CORRECT DIRECTORY PATHS BELOW

import os # type:ignore # isort:skip # fmt:skip # noqa # nopep8
import sys # type:ignore # isort:skip # fmt:skip # noqa # nopep8
from pathlib import Path # type:ignore # isort:skip # fmt:skip # noqa # nopep8

mod = sys.modules[__name__]

code_dir = None
code_dir_name = 'Code'
unwanted_subdir_name = 'Analysis'

if code_dir_name not in str(Path.cwd()).split('/')[-1]:
    for _ in range(5):

        parent_path = str(Path.cwd().parents[_]).split('/')[-1]

        if (code_dir_name in parent_path) and (unwanted_subdir_name not in parent_path):

            code_dir = str(Path.cwd().parents[_])

            if code_dir is not None:
                break
else:
    code_dir = str(Path.cwd())
sys.path.append(code_dir)

# %load_ext autoreload
# %autoreload 2


from setup_module.imports import * # type:ignore # isort:skip # fmt:skip # noqa # nopep8


df_jobs = pd.read_pickle(f'{df_save_dir}df_raw.pkl')


df_jobs['Job Description'] = df_jobs['Job Description'].progress_apply(lambda x: unicodedata.normalize('NFKD', x.encode('ascii', 'ignore').decode('utf-8', 'ignore')))


# len = 204113
len(df_jobs)


df_jobs.info()


df_jobs.columns


# Clean columns
df_jobs.columns = df_jobs.columns.to_series().progress_apply(lambda x: str(x).strip())


# Missing values: Job Description = 15, Job ID = 0
df_jobs.isna().sum()


# Drop NA
df_jobs = df_jobs.dropna(axis='index', how='all')
df_jobs = df_jobs.dropna(axis='columns', how='all')
df_jobs['Job Description'] = df_jobs['Job Description'].progress_apply(lambda x: unicodedata.normalize('NFKD', x.encode('ascii', 'ignore').decode('utf-8', 'ignore')))


# len = 204085
len(df_jobs)


df_jobs['Job Description'] = df_jobs['Job Description'].astype(str)


df_jobs['Job Description'] = df_jobs['Job Description'].progress_apply(lambda x: x.strip())


df_jobs['Job Description'] = df_jobs[
    'Job Description'
].progress_apply(lambda x: x if isinstance(x, str) else ast.literal_eval(x))
df_jobs['Job Description'] = df_jobs['Job Description'].progress_apply(lambda x: unicodedata.normalize('NFKD', x.encode('ascii', 'ignore').decode('utf-8', 'ignore')))


# Drop duplicates on subset of 'Job Description'
df_jobs = df_jobs.drop_duplicates(subset=['Job Description'], keep='first')


# Conver Job ID and Sentence to str
str_cols = [
    'Job ID',
    'Job Description',
]

for col in str_cols:
    df_jobs[col] = df_jobs[col].astype(str)
    print(f'{col} converted to str.' if all(df_jobs[col].progress_apply(lambda x: isinstance(x, str))) else f'{col} NOT converted to str.')


# Drop duplicates on subset of 'Job Description'
df_jobs = df_jobs.drop_duplicates(subset=['Job ID', 'Job Description'], keep='first', ignore_index=True)


# len = 62579
len(df_jobs)


# len = 62577
len(df_jobs)


# Save df_jobs to file
assert len(df_jobs) > 0 and isinstance(df_jobs, pd.DataFrame), f'ERORR: LENGTH OF DF = {len(df_jobs)}'
df_jobs.to_pickle(f'{df_save_dir}df_jobs_raw_dropped.pkl')
df_jobs.to_csv(f'{df_save_dir}df_jobs_raw_dropped.csv', index=False)


# Detect job description language

### START HERE IF SOURCING FROM DF_JOBS_RAW_DROPPED
### PLEASE SET CORRECT DIRECTORY PATHS BELOW


import os # type:ignore # isort:skip # fmt:skip # noqa # nopep8
import sys # type:ignore # isort:skip # fmt:skip # noqa # nopep8
from pathlib import Path # type:ignore # isort:skip # fmt:skip # noqa # nopep8

mod = sys.modules[__name__]

code_dir = None
code_dir_name = 'Code'
unwanted_subdir_name = 'Analysis'

if code_dir_name not in str(Path.cwd()).split('/')[-1]:
    for _ in range(5):

        parent_path = str(Path.cwd().parents[_]).split('/')[-1]

        if (code_dir_name in parent_path) and (unwanted_subdir_name not in parent_path):

            code_dir = str(Path.cwd().parents[_])

            if code_dir is not None:
                break
else:
    code_dir = str(Path.cwd())
sys.path.append(code_dir)

# %load_ext autoreload
# %autoreload 2


from setup_module.imports import *  # type:ignore # isort:skip # fmt:skip # noqa # nopep8


df_jobs = pd.read_pickle(f'{df_save_dir}df_raw_dropped.pkl')


# 62577
len(df_jobs)


df_jobs.info()


# This is a manually collected dictionary of incorrect/faulty keywords in scraped site data
with open(f'{scraped_data}CBS/Data/keyword_trans_dict.txt') as f:
    keyword_trans_dict = json.load(f)

if len(df_jobs['Search Keyword'].loc[df_jobs['Search Keyword'].isin(list(keyword_trans_dict.keys()))]) != 0:
    df_jobs = fix_keywords(df_jobs)


%%time
translator = Translator()
googletrans_readtime_error = googletrans.client.httpx._client.httpcore._exceptions.ReadTimeout

if 'Language' not in df_jobs.columns:
    df_jobs['Language'] = np.nan # create Language col and fill it with nan

# try:
#     time.sleep(60)
#     df_jobs['Language'] = df_jobs['Job Description'].progress_apply(lambda x: translator.detect(str(x).lower().strip()).lang)
# except:
#     time.sleep(3600)
#     df_jobs['Language'] = df_jobs['Job Description'].progress_apply(lambda x: translator.detect(str(x).lower().strip()).lang)

for idx, row in df_jobs.iterrows():
    # This part ensures we don't start lang detection from index 0 if some lang detection was already done
    if len(str(row['Job Description'])) != 0:
        if type(row['Language']) == float and np.isnan(row['Language']): #if lang is nan, detect language

            try:
                print(f'Row {idx}: Language detection in progress.')
#                 time.sleep(10)
                df_jobs.loc[idx, 'Language'] = str(translator.detect(str(row['Job Description']).lower().strip()).lang)
                print(f'Row {idx}: Language detection done.')
            except:# googletrans_readtime_error:

                if len(df_jobs) > 0 and isinstance(df_jobs, pd.DataFrame):
                    df_jobs.to_pickle(f'{df_save_dir}df_raw_language_detected.pkl')

                    df_jobs.to_csv(f'{df_save_dir}df_raw_language_detected.csv', index=False)
                else:
                    print(f'ERORR: LENGTH OF DF = {len(df_jobs)}')

                print(f'Row {idx}: Sleeping for half an hour starting at {time.strftime("%I:%M:%S %p", time.localtime())}.')
                print('-'*30)
                time.sleep(1800)
                print(f'Row {idx}: Done sleeping.')
                print('-'*30)
                print(f'Row {idx}: Language detection in progress.')
                df_jobs.loc[idx, 'Language'] = str(translator.detect(str(row['Job Description']).lower().strip()).lang)
                print(f'Row {idx}: Language detection done.')

        else: # elif lang is not nan, skip and go to next idx
            continue

if len(df_jobs) > 0 and isinstance(df_jobs, pd.DataFrame):
    df_jobs.to_pickle(f'{df_save_dir}df_raw_language_detected.pkl')

    df_jobs.to_csv(f'{df_save_dir}df_raw_language_detected.csv', index=False)
else:
    print(f'ERORR: LENGTH OF DF = {len(df_jobs)}')


# nl = 44863, en = 17591, ['en', 'nl'] = 8
df_jobs['Language'].value_counts()


assert len(df_jobs) > 0 and isinstance(df_jobs, pd.DataFrame), f'ERORR: LENGTH OF DF = {len(df_jobs)}'
df_jobs.to_pickle(f'{df_save_dir}df_jobs_raw_language_detected.pkl')
df_jobs.to_csv(f'{df_save_dir}df_jobs_raw_language_detected.csv', index=False)
